+++ Controllers/RoomController.php
@@ -3,6 +3,7 @@
 
     use DAO\RoomDAO as RoomDAO;
     use DAO\CinemaDAO as CinemaDAO;
+    use DAO\FunctionDAO as FunctionDAO;
     use Models\Room as Room;
     use Models\Cinema as Cinema;
     //use Models\Function as Function;
@@ -11,11 +12,13 @@
     {
         private $roomDAO;
         private $cinemaDAO;
+        private $functionDAO;
 
         public function __construct()
         {
             $this->roomDAO = new RoomDAO();
             $this->cinemaDAO = new CinemaDAO();
+            $this->functionDAO = new FunctionDAO();
         }
 
         public function ShowAddView($idCinema)
@@ -51,10 +54,17 @@
 
         public function Remove($idRoom)
         {
-            $room = $this->roomDAO->GetById($idRoom);//LO HAGO PARA PODERTENER EL ID DEL CINE Y VOLVER A MOSTRAR LA LISTA
+            $room = $this->roomDAO->GetById($idRoom);//LO HAGO PARA PODER TENER EL ID DEL CINE Y VOLVER A MOSTRAR LA LISTA
             //Trabajamos con baja logica para seguir teniendo persistencia de todo
-            $this->roomDAO->Remove($idRoom);//VER SI PASO EL ROOM O SOLO ID
             
+            if(empty($this->functionDAO->Get($room)))//SI LA SALA NO TIENE FUNCIONES, la puedo borrar
+            {
+                $this->roomDAO->Remove($idRoom);//VER SI PASO EL ROOM O SOLO ID
+                //AGREGAR SCRIPT DE EXITO
+            }
+            else {}//SCRIPT DE ERROR; POR TENER FUNCIONES}
+            //Si borro una sala que tiene funciones,no me tiene que dejar.
+            // o preguntar que quiero hacer y borrar las funciones. ver luego tema de entradas vendidas
             $this->ShowListView($room->getCinema()->getId());
         }
